refactor: separate ScheduleSpareTimeField and ScheduleSpareTimeForm into different files
