+++ lib/presentation/onboarding/schedule_spare_time/screens/schedule_spare_time_form.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:on_time_front/presentation/onboarding/components/onboarding_page_view_layout.dart';
+import 'package:on_time_front/presentation/onboarding/schedule_spare_time/components/shcedule_spare_time_field.dart';
 import 'package:on_time_front/presentation/onboarding/schedule_spare_time/cubit/schedule_spare_time_cubit.dart';
-import 'package:on_time_front/presentation/shared/components/time_stepper.dart';
 
 class ScheduleSpareTimeForm extends StatefulWidget {
   const ScheduleSpareTimeForm({
@@ -42,39 +42,6 @@ class _ScheduleSpareTimeFormState extends State<ScheduleSpareTimeForm> {
             },
           );
         },
-      ),
-    );
-  }
-}
-
-class ScheduleSpareTimeField extends StatelessWidget {
-  const ScheduleSpareTimeField({
-    super.key,
-    required this.lowerBound,
-    required this.spareTime,
-    required this.onSpareTimeDecreased,
-    required this.onSpareTimeIncreased,
-  });
-
-  final Duration spareTime;
-  final Duration lowerBound;
-  final VoidCallback onSpareTimeIncreased;
-  final VoidCallback onSpareTimeDecreased;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.only(top: 90.0),
-      child: Column(
-        children: [
-          TimeStepper(
-            onSpareTimeIncreased: onSpareTimeIncreased,
-            onSpareTimeDecreased: onSpareTimeDecreased,
-            lowerBound: lowerBound,
-            value: spareTime,
-          ),
-          Expanded(child: SizedBox()),
-        ],
       ),
     );
   }
